refactor: create use search souvenir hooks
